+++ src/fhir/common/sdc-export-common.js
@@ -1081,10 +1081,12 @@ function addCommonSDCExportFns(ns) {
       // extractable and some codes have ObsExtract=false, the item is still
       // extractable (extracts all codes except those with ObsExtract=false). See
       // https://chat.fhir.org/#narrow/channel/179255-questionnaire/topic/Observation.20extraction.20with.20both.20item.20and.20item.2Ecode/with/517623297.
-      return true;
+      item._extractValue = true;
     } else {
-      return self._getExtractedObsCodes(item);
+      item._extractValue = self._getExtractedObsCodes(item);
     }
+    self._findObsExtractValueCodeParent(item);
+    return item._extractValue;
   };
 
 
@@ -1155,6 +1157,28 @@ function addCommonSDCExportFns(ns) {
     // should be extracted. In this case, ObsExtract=true extension is not
     // necessary on item level.
     return hasCodeLevelObsExtractTrue;
+  };
+
+
+  /**
+   * If the item has ObsExtract valueCode, find the nearest extractable parent and
+   * set _obsExtractParentLinkId on the item to the parent linkId so we will know which
+   * extracted parent Observation to link to.
+   * @param item an item in Questionnaire
+   */
+  self._findObsExtractValueCodeParent = function (item) {
+    const obsExtractValueCode = item._fhirExt && item._fhirExt[this.fhirExtObsExtract] && item._fhirExt[this.fhirExtObsExtract][0].valueCode;
+    if (obsExtractValueCode) {
+      let currentItem = item._parentItem;
+      while (currentItem) {
+        if (currentItem._extractValue) {
+          item._obsExtractValueCode = obsExtractValueCode;
+          item._obsExtractParentLinkId = currentItem.linkId;
+          return;
+        }
+        currentItem = currentItem._parentItem;
+      }
+    }
   };