+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [38.2.0] 2025-08-08
+### Added
+- Tab layout support.
+
+## [38.1.0] 2025-08-07
+### Added
+- Support for observationExtract valueCode relationships (component or member).
+
+## [38.0.2] 2025-07-31
+### Fixed
+- Child questions were also highlighted with background color
+  when parent question was focused.
+
 ## [38.0.1] 2025-07-24
 ### Fixed
 - Some languages use a different decimal character like "," instead
@@ -28,7 +41,7 @@ This project follows [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fixed a bug that the matrix layout ('table' in itemControl extension) 
   does not always display answers loaded from answerValueSet.
-  
+
 ## [36.20.2] 2025-07-09
 ### Fixed
 - answerValueSet expansion error under the field should still show